+++ app/utils/flowtype.js
@@ -74,7 +74,7 @@ export type selectedSearchState = {
         parameters:DatasourceParameter,
         dates: {
             available: {start:Date, end:Date},
-            selected: {start:Date, end:Date},
+            selected: {start:?Date, end:?Date},
         }
     };
 export type sensorsState= {