+++ CHANGELOG.md
@@ -80,15 +80,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-963](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-963)
 - Make the subtitle in cards for optional on the Trends and Analysis Pages
   [GEOD-961](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-961)
+- Removed all instances of 'UPDATE_TRENDS_REGIONS_POINTS' and associated code
+  [GEOD-975](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-975)
 - Add the ability for Detail Links in V3 to go to V2
   [GEOD-959](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-959)
 
 ### Fixed
 - Bug in filtering for custom locations
   [GEOD-925](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-925)
-- Removed the card max-height on the V3 Search Page
-  [GEOD-982](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-982)
-- Fixed bug with drawing shapes not erasing when location card does not exist
-  [GEOD-981](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-981)
+- Bug in selected search not being updated when a filter was removed
+  [GEOD-953](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-953)
 - Fixed bug with Draw location filter not cleared up
   [GEOD-968](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-968)
+- Fixed bug with drawing shapes not erasing when location card does not exist
+  [GEOD-981](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-981)
+- Removed the card max-height on the V3 Search Page
+  [GEOD-982](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-982)
+
+## Changed
+- Changed how the Circle is created, and how the associated coordinates are assembled
+  [GEOD-965](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-965)
+